scripts/requirements.txt
edgartools>=3.0.0
markdownify>=0.14.1
